whois 5.0.17
This is the whois used by Debian and its down-stream distributions.

Default whois on OSX doesn't offer as many features. I've shadowed the
system whois because I can't think that any scripts would be using
whois programmatically or that anything would really depend on this,
so if you've brew installed it, you probably want it to be the
default.